refactor: move IconWithPopover to atoms
