+++ src/components/atoms/index.ts
@@ -4,6 +4,7 @@ export {default as Dots} from './Dots';
 export {default as FileExplorer} from './FileExplorer';
 export {default as HotkeyLabel} from './HotkeyLabel';
 export {default as Icon} from './Icon';
+export {default as IconWithPopover} from './IconWithPopover';
 export {default as ImageIcon} from './ImageIcon';
 export {default as InputTags} from './InputTags';
 export {default as KeyValueInput} from './KeyValueInput';
